requirements.txt
Flask
python-dotenv
black
pypika
simplejson
flask-cors
pymysql

watchdog
blinker==1.4
Flask-HTTPAuth==3.1.1
Flask-HTTPS==0.1.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
MarkupSafe==0.23
SQLAlchemy==1.0.12
